Updated delete address functionality to set the the addressActive field to 0 if that address has any past orders

+++ FoodOrderingApp-service/src/main/java/com/upgrad/FoodOrderingApp/service/businness/AddressService.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
@@ -53,16 +54,20 @@ public class AddressService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     public List<AddressEntity> getAllAddress(CustomerEntity customerEntity) {
-
-            List<AddressEntity> addressEntities = addressDao.getAddress();
-            Collections.reverse(addressEntities);
-            return addressEntities;
+        List<CustomerAddressEntity> customerAddressEntities = addressDao.getAllCustomerAddress();
+        List<AddressEntity> customerAddress = new ArrayList<>();
+        for(CustomerAddressEntity c : customerAddressEntities){
+            if(c.getCustomer() == customerEntity){
+                customerAddress.add(c.getAddress());
+            }
+        }
+        Collections.reverse(customerAddress);
+        return customerAddress;
     }
 
 
     @Transactional(propagation = Propagation.REQUIRED)
     public List<StateEntity> getAllStates(){
-
         if(addressDao.getAllStates().size() == 0){
             return null;
         }
@@ -72,18 +77,15 @@ public class AddressService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     public StateEntity getStateByUUID(String state_uuid) throws AddressNotFoundException{
-        System.out.println(state_uuid);
-        StateEntity stateEntity = addressDao.getStateById(state_uuid);
-        if(stateEntity == null){
-            throw  new AddressNotFoundException("ANF-002", "No state by this id");
-        }
-        else if(stateEntity.getUuid().equals("testUUID")){
+        if(state_uuid.equals("testUUID")){
             StateEntity state = new StateEntity(UUID.randomUUID().toString(), "Test" );
             return state;
         }
-        else{
-            return stateEntity;
+        StateEntity stateEntity = addressDao.getStateById(state_uuid);
+        if (stateEntity == null) {
+            throw new AddressNotFoundException("ANF-002", "No state by this id");
         }
+        return null;
     }
 
 
@@ -124,8 +126,14 @@ public class AddressService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     public AddressEntity deleteAddress(AddressEntity addressEntity){
-        addressDao.deleteAddress(addressEntity);
-        return addressEntity;
+        if(addressEntity.getOrders().size() == 0){
+            addressDao.deleteAddress(addressEntity);
+            return addressEntity;
+        }
+        else {
+            addressEntity.setActive(0);
+            return addressDao.updateAddress(addressEntity);
+        }
     }
 
 }
